Expand Dockerfile examples
If we invoke the 'assemble' script during docker build, the playbook
should be imported first so that e.g. a requirements.txt in the repo is
copied. Also, in this case we must copy to /tmp/src instead of APP_HOME,
as this is where 'assemble' expects it.

As this starts to complicate the example Dockerfile, creating 2 examples
instead: a real simple "just package my playbook" and an advanced version that
explains how to install dependencies